+++ packages/plugin-text/src/factory/types.ts
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { OverlayContextValue } from '@edtr-io/core'
+import { OverlayContextValue, Plugin, PluginState } from '@edtr-io/core'
 import { TextPlugin } from '..'
 import { SlateEditorAdditionalProps } from './editor'
 
@@ -12,4 +12,6 @@ export type SlatePluginClosure = React.RefObject<{
   overlayContext: OverlayContextValue
   name: string
   parent?: SlateEditorAdditionalProps
+  replace?: (options?: PluginState) => void
+  plugins: Record<string, Plugin>
 }>