+++ scripts/changelog.ts
@@ -216,7 +216,8 @@ async function exec(): Promise<void> {
     {
       added: [
         '**plugin-files**. Add plugin for file uploads',
-        '**plugin-rows**. Pass `renderIntoExtendedSettings` and `PrimarySettings` as props to children'
+        '**plugin-rows**. Pass `renderIntoExtendedSettings` and `PrimarySettings` as props to children',
+        '**plugin-text**. Poweruser-Feature: Add new plugins with /[plugintitle] in editmode'
       ],
       changed: [
         '**plugin-rows**. Adapt styles and controls implemented by schul-cloud, including drag&drop',